+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 3.2.6 (April 28, 2026)
 * Add support for EU Soverign Cloud region
 
+ENHANCEMENTS:
+* Added `disallow_weak_passwords` field to `sumologic_password_policy` resource to support the Disallow Weak Passwords setting introduced in the Sumo Logic password policy API.
+
 DOCS:
 * Added documentation for `direction` field in `logs_anomaly_condition` and `metrics_anomaly_condition` for `sumologic_monitor` resource
 * Added Sumo help docs link for `sumologic_event_extraction_rule` resource documentation.